Update componentize-go build command and rename package
