+++ examples/redis/spin.toml
@@ -16,5 +16,4 @@ component = "echo-message"
 [component.echo-message]
 source = "main.wasm"
 [component.echo-message.build]
-command = "go tool componentize-go --world fermyon:spin/redis-trigger build"
-
+command = "go tool componentize-go --world spin:up/redis-trigger build"
