+++ redis/redis.go
@@ -7,7 +7,7 @@ import (
 	"fmt"
 
 	_ "github.com/spinframework/spin-go-sdk/v3/exports/fermyon_spin_redis_trigger/wit_exports"
-	incominghandler "github.com/spinframework/spin-go-sdk/v3/exports/spin_up_redis_trigger_4_0_0/export_spin_up_inbound_redis"
+	incominghandler "github.com/spinframework/spin-go-sdk/v3/exports/spin_up_redis_trigger_4_0_0/export_spin_redis_3_0_0_inbound_redis"
 	redis "github.com/spinframework/spin-go-sdk/v3/imports/spin_redis_3_0_0_redis"
 	wit "go.bytecodealliance.org/pkg/wit/types"
 )
